restructured code base
