+++ src/whir/mod.rs
@@ -1,4 +1,3 @@
-pub mod batch_proof;
 pub mod committer;
 pub mod proof;
 pub mod prover;
@@ -378,6 +377,493 @@ mod test {
                 0,
                 1,
                 SumcheckStrategy::default(),
+            );
+        }
+    }
+}
+
+#[cfg(test)]
+mod batch_tests {
+    use alloc::vec;
+
+    use p3_baby_bear::{BabyBear, Poseidon2BabyBear};
+    use p3_challenger::{DuplexChallenger, FieldChallenger};
+    use p3_field::{Field, PrimeCharacteristicRing, extension::BinomialExtensionField};
+    use p3_merkle_tree::MerkleTreeMmcs;
+    use p3_multilinear_util::{evals::EvaluationsList, multilinear::MultilinearPoint};
+    use p3_symmetric::{PaddingFreeSponge, TruncatedPermutation};
+    use rand::{RngExt, SeedableRng, rngs::SmallRng};
+
+    use crate::{
+        constraints::statement::EqStatement,
+        parameters::{FoldingFactor, ProtocolParameters, SecurityAssumption, WhirConfig},
+        sumcheck::SumcheckData,
+        whir::{
+            committer::{
+                reader::BatchCommitmentReader,
+                writer::BatchCommitmentWriter,
+            },
+            proof::{BatchWhirProof, WhirProof},
+            prover::Prover,
+            verifier::Verifier,
+        },
+    };
+
+    type F = BabyBear;
+    type EF = BinomialExtensionField<F, 4>;
+    type Perm = Poseidon2BabyBear<16>;
+    type MyHash = PaddingFreeSponge<Perm, 16, 8, 8>;
+    type MyCompress = TruncatedPermutation<Perm, 2, 8, 16>;
+    type MyChallenger = DuplexChallenger<F, Perm, 16, 8>;
+    type PackedF = <F as Field>::Packing;
+    type MyMmcs = MerkleTreeMmcs<PackedF, PackedF, MyHash, MyCompress, 2, 8>;
+
+    fn make_batch_test_config(
+        num_variables: usize,
+        folding_factor: FoldingFactor,
+    ) -> (
+        WhirConfig<EF, F, MyMmcs, MyChallenger>,
+        ProtocolParameters<MyMmcs>,
+        SmallRng,
+    ) {
+        let mut rng = SmallRng::seed_from_u64(42);
+        let perm = Perm::new_from_rng_128(&mut rng);
+        let mmcs = MyMmcs::new(MyHash::new(perm.clone()), MyCompress::new(perm), 0);
+
+        let whir_params = ProtocolParameters {
+            security_level: 32,
+            pow_bits: 0,
+            rs_domain_initial_reduction_factor: 1,
+            folding_factor,
+            mmcs,
+            soundness_type: SecurityAssumption::UniqueDecoding,
+            starting_log_inv_rate: 1,
+        };
+
+        let params =
+            WhirConfig::<EF, F, MyMmcs, MyChallenger>::new(num_variables, whir_params.clone());
+
+        (params, whir_params, rng)
+    }
+
+    #[test]
+    fn test_selector_round_correctness() {
+        let (params, _whir_params, mut rng) = make_batch_test_config(6, FoldingFactor::Constant(3));
+
+        let num_variables = 6;
+        let num_evals = 1 << num_variables;
+        let poly_a = EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+        let poly_b = EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+
+        let z_a = MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+        let z_b = MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+        let v_a: EF = poly_a.evaluate_hypercube_base(&z_a);
+        let v_b: EF = poly_b.evaluate_hypercube_base(&z_b);
+
+        let perm = Perm::new_from_rng_128(&mut rng);
+        let mut challenger = MyChallenger::new(perm);
+        let alpha: EF = challenger.sample_algebra_element();
+
+        let mut selector_data: SumcheckData<F, EF> = SumcheckData::default();
+
+        let prover = Prover(&params);
+        let (sumcheck_prover, r_0) = prover.selector_round(
+            &mut selector_data,
+            &mut challenger,
+            &poly_a,
+            &poly_b,
+            &z_a,
+            &z_b,
+            v_a,
+            v_b,
+            alpha,
+        );
+
+        // Verify: g should equal r_0·f_a + (1-r_0)·f_b at every point
+        let g = sumcheck_prover.evals();
+        let one_minus_r0 = EF::ONE - r_0;
+        for i in 0..num_evals {
+            let expected = r_0 * EF::from(poly_a.as_slice()[i])
+                + one_minus_r0 * EF::from(poly_b.as_slice()[i]);
+            assert_eq!(g.as_slice()[i], expected, "g mismatch at index {i}");
+        }
+
+        // Verify: selector sumcheck stored exactly 1 round of data
+        assert_eq!(selector_data.polynomial_evaluations().len(), 1);
+
+        // Verify: c0 == α·v_b (the h(0) value)
+        let [c0, _c2] = selector_data.polynomial_evaluations()[0];
+        assert_eq!(c0, alpha * v_b);
+    }
+
+    /// Run the full batch prove + verify flow and check transcript consistency.
+    fn run_batch_prove_verify(num_variables: usize, folding_factor: FoldingFactor) {
+        let (params, whir_params, mut rng) = make_batch_test_config(num_variables, folding_factor);
+        let dft = p3_dft::Radix2DFTSmallBatch::<F>::default();
+
+        let num_evals = 1 << num_variables;
+        let poly_a = EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+        let poly_b = EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+
+        let z_a = MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+        let z_b = MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+        let v_a: EF = poly_a.evaluate_hypercube_base(&z_a);
+        let v_b: EF = poly_b.evaluate_hypercube_base(&z_b);
+
+        let mut statement_a = EqStatement::initialize(num_variables);
+        statement_a.add_evaluated_constraint(z_a, v_a);
+        let mut statement_b = EqStatement::initialize(num_variables);
+        statement_b.add_evaluated_constraint(z_b, v_b);
+
+        // Both challengers start from the same state
+        let perm = Perm::new_from_rng_128(&mut rng);
+        let base_challenger = MyChallenger::new(perm);
+        let mut prover_challenger = base_challenger.clone();
+        let mut verifier_challenger = base_challenger;
+
+        // --- Prover side ---
+        let batch_committer = BatchCommitmentWriter::new(&params);
+        let mut proof = BatchWhirProof {
+            commitment_a: None,
+            commitment_b: None,
+            initial_ood_answers: [vec![], vec![]],
+            selector_sumcheck: SumcheckData::default(),
+            inner_proof: WhirProof::<F, EF, MyMmcs>::from_protocol_parameters(
+                &whir_params,
+                num_variables,
+            ),
+        };
+
+        let batch_data = batch_committer.commit(
+            &dft,
+            &mut proof,
+            &mut prover_challenger,
+            &poly_a,
+            &poly_b,
+        );
+
+        let prover = Prover(&params);
+        prover
+            .batch_prove(
+                &dft,
+                &mut proof,
+                &mut prover_challenger,
+                batch_data.prover_data_a,
+                batch_data.prover_data_b,
+                &poly_a,
+                &poly_b,
+                &statement_a,
+                &statement_b,
+                &batch_data.ood_statement_a,
+                &batch_data.ood_statement_b,
+            )
+            .unwrap();
+
+        let checkpoint_prover: EF = prover_challenger.sample_algebra_element();
+
+        // --- Verifier side ---
+        let batch_reader = BatchCommitmentReader::new(&params);
+        let (parsed_a, parsed_b) =
+            batch_reader.parse_batch_commitment(&proof, &mut verifier_challenger);
+
+        let verifier = Verifier::new(&params);
+        let (_folding_randomness, _r_0) = verifier
+            .batch_verify(
+                &proof,
+                &mut verifier_challenger,
+                &parsed_a,
+                &parsed_b,
+                &statement_a,
+                &statement_b,
+            )
+            .unwrap();
+
+        let checkpoint_verifier: EF = verifier_challenger.sample_algebra_element();
+        assert_eq!(checkpoint_prover, checkpoint_verifier);
+    }
+
+    #[test]
+    fn test_batch_prove_verify_folding_4() {
+        run_batch_prove_verify(10, FoldingFactor::Constant(4));
+    }
+
+    #[test]
+    fn test_batch_prove_verify_folding_2() {
+        run_batch_prove_verify(8, FoldingFactor::Constant(2));
+    }
+
+    #[test]
+    fn test_batch_prove_verify_folding_3() {
+        run_batch_prove_verify(9, FoldingFactor::Constant(3));
+    }
+
+    #[test]
+    fn test_batch_prove_verify_mixed_folding() {
+        run_batch_prove_verify(10, FoldingFactor::ConstantFromSecondRound(4, 2));
+    }
+
+    #[cfg(not(debug_assertions))]
+    mod performance {
+        extern crate std;
+
+        use std::time::Instant;
+
+        use super::*;
+        use crate::{
+            parameters::SumcheckStrategy,
+            whir::committer::writer::CommitmentWriter,
+        };
+
+        /// Run a single-polynomial commit + prove cycle. Returns elapsed time.
+        #[allow(clippy::too_many_arguments)]
+        fn single_poly_prove(
+            params: &WhirConfig<EF, F, MyMmcs, MyChallenger>,
+            whir_params: &ProtocolParameters<MyMmcs>,
+            dft: &p3_dft::Radix2DFTSmallBatch<F>,
+            poly: EvaluationsList<F>,
+            num_variables: usize,
+            num_points: usize,
+            rng: &mut SmallRng,
+        ) -> std::time::Duration {
+            let mut statement = params.initial_statement(poly, SumcheckStrategy::default());
+            for _ in 0..num_points {
+                let point =
+                    MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+                let _ = statement.evaluate(&point);
+            }
+
+            let perm = Perm::new_from_rng_128(rng);
+            let mut challenger = MyChallenger::new(perm);
+
+            let committer = CommitmentWriter::new(params);
+            let mut proof = WhirProof::<F, EF, MyMmcs>::from_protocol_parameters(
+                whir_params,
+                num_variables,
+            );
+
+            let start = Instant::now();
+
+            let prover_data = committer
+                .commit(dft, &mut proof, &mut challenger, &mut statement)
+                .unwrap();
+
+            let prover = Prover(params);
+            prover
+                .prove(dft, &mut proof, &mut challenger, &statement, prover_data)
+                .unwrap();
+
+            start.elapsed()
+        }
+
+        /// Run a batch prove cycle for two polynomials. Returns elapsed time.
+        #[allow(clippy::too_many_arguments)]
+        fn batch_poly_prove(
+            params: &WhirConfig<EF, F, MyMmcs, MyChallenger>,
+            whir_params: &ProtocolParameters<MyMmcs>,
+            dft: &p3_dft::Radix2DFTSmallBatch<F>,
+            poly_a: &EvaluationsList<F>,
+            poly_b: &EvaluationsList<F>,
+            num_variables: usize,
+            rng: &mut SmallRng,
+        ) -> std::time::Duration {
+            let z_a =
+                MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+            let z_b =
+                MultilinearPoint::expand_from_univariate(rng.random(), num_variables);
+            let v_a: EF = poly_a.evaluate_hypercube_base(&z_a);
+            let v_b: EF = poly_b.evaluate_hypercube_base(&z_b);
+
+            let mut statement_a = EqStatement::initialize(num_variables);
+            statement_a.add_evaluated_constraint(z_a, v_a);
+            let mut statement_b = EqStatement::initialize(num_variables);
+            statement_b.add_evaluated_constraint(z_b, v_b);
+
+            let perm = Perm::new_from_rng_128(rng);
+            let mut challenger = MyChallenger::new(perm);
+
+            let batch_committer = BatchCommitmentWriter::new(params);
+            let mut proof = BatchWhirProof {
+                commitment_a: None,
+                commitment_b: None,
+                initial_ood_answers: [vec![], vec![]],
+                selector_sumcheck: SumcheckData::default(),
+                inner_proof: WhirProof::<F, EF, MyMmcs>::from_protocol_parameters(
+                    whir_params,
+                    num_variables,
+                ),
+            };
+
+            let start = Instant::now();
+
+            let batch_data = batch_committer.commit(
+                dft,
+                &mut proof,
+                &mut challenger,
+                poly_a,
+                poly_b,
+            );
+
+            let prover = Prover(params);
+            prover
+                .batch_prove(
+                    dft,
+                    &mut proof,
+                    &mut challenger,
+                    batch_data.prover_data_a,
+                    batch_data.prover_data_b,
+                    poly_a,
+                    poly_b,
+                    &statement_a,
+                    &statement_b,
+                    &batch_data.ood_statement_a,
+                    &batch_data.ood_statement_b,
+                )
+                .unwrap();
+
+            start.elapsed()
+        }
+
+        #[test]
+        fn test_batch_vs_separate_performance() {
+            const ITERS: usize = 5;
+
+            let num_variables = 16;
+            let folding_factor = FoldingFactor::Constant(4);
+            let (params, whir_params, mut rng) =
+                make_batch_test_config(num_variables, folding_factor);
+            let dft = p3_dft::Radix2DFTSmallBatch::<F>::default();
+
+            let num_evals = 1 << num_variables;
+            let poly_a =
+                EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+            let poly_b =
+                EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+            let mut separate_total = std::time::Duration::ZERO;
+            for _ in 0..ITERS {
+                let t1 = single_poly_prove(
+                    &params,
+                    &whir_params,
+                    &dft,
+                    poly_a.clone(),
+                    num_variables,
+                    1,
+                    &mut rng,
+                );
+                let t2 = single_poly_prove(
+                    &params,
+                    &whir_params,
+                    &dft,
+                    poly_b.clone(),
+                    num_variables,
+                    1,
+                    &mut rng,
+                );
+                separate_total += t1 + t2;
+            }
+
+            let mut batch_total = std::time::Duration::ZERO;
+            for _ in 0..ITERS {
+                let t = batch_poly_prove(
+                    &params,
+                    &whir_params,
+                    &dft,
+                    &poly_a,
+                    &poly_b,
+                    num_variables,
+                    &mut rng,
+                );
+                batch_total += t;
+            }
+
+            let separate_avg = separate_total / ITERS as u32;
+            let batch_avg = batch_total / ITERS as u32;
+            let speedup = separate_avg.as_secs_f64() / batch_avg.as_secs_f64();
+
+            std::eprintln!();
+            std::eprintln!(
+                "=== Batch vs Separate Performance ({num_variables} variables, {ITERS} iterations) ==="
+            );
+            std::eprintln!("  Two separate proofs: {separate_avg:?} avg");
+            std::eprintln!("  One batch proof:     {batch_avg:?} avg");
+            std::eprintln!("  Speedup:             {speedup:.2}x");
+            std::eprintln!();
+
+            // The batch proof should be faster than two separate proofs.
+            assert!(
+                batch_avg < separate_avg,
+                "Batch proof ({batch_avg:?}) should be faster than two separate proofs ({separate_avg:?})"
+            );
+        }
+
+        #[test]
+        fn test_batch_vs_separate_performance_large() {
+            const ITERS: usize = 3;
+
+            let num_variables = 20;
+            let folding_factor = FoldingFactor::Constant(4);
+            let (params, whir_params, mut rng) =
+                make_batch_test_config(num_variables, folding_factor);
+            let dft = p3_dft::Radix2DFTSmallBatch::<F>::default();
+
+            let num_evals = 1 << num_variables;
+            let poly_a =
+                EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+            let poly_b =
+                EvaluationsList::new((0..num_evals).map(|_| rng.random()).collect());
+
+            let mut separate_total = std::time::Duration::ZERO;
+            for _ in 0..ITERS {
+                let t1 = single_poly_prove(
+                    &params,
+                    &whir_params,
+                    &dft,
+                    poly_a.clone(),
+                    num_variables,
+                    1,
+                    &mut rng,
+                );
+                let t2 = single_poly_prove(
+                    &params,
+                    &whir_params,
+                    &dft,
+                    poly_b.clone(),
+                    num_variables,
+                    1,
+                    &mut rng,
+                );
+                separate_total += t1 + t2;
+            }
+
+            let mut batch_total = std::time::Duration::ZERO;
+            for _ in 0..ITERS {
+                let t = batch_poly_prove(
+                    &params,
+                    &whir_params,
+                    &dft,
+                    &poly_a,
+                    &poly_b,
+                    num_variables,
+                    &mut rng,
+                );
+                batch_total += t;
+            }
+
+            let separate_avg = separate_total / ITERS as u32;
+            let batch_avg = batch_total / ITERS as u32;
+            let speedup = separate_avg.as_secs_f64() / batch_avg.as_secs_f64();
+
+            std::eprintln!();
+            std::eprintln!(
+                "=== Batch vs Separate Performance ({num_variables} variables, {ITERS} iterations) ==="
+            );
+            std::eprintln!("  Two separate proofs: {separate_avg:?} avg");
+            std::eprintln!("  One batch proof:     {batch_avg:?} avg");
+            std::eprintln!("  Speedup:             {speedup:.2}x");
+            std::eprintln!();
+
+            assert!(
+                batch_avg < separate_avg,
+                "Batch proof ({batch_avg:?}) should be faster than two separate proofs ({separate_avg:?})"
             );
         }
     }
